Remove redundant files and refactor Ui.java into UiPrinter.java
