+++ src/main/java/duke/controller/MainWindow.java
@@ -50,6 +50,13 @@ public class MainWindow extends AnchorPane {
     @FXML
     public void initialize() {
         scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+        greet();
+    }
+
+    private void greet() {
+        DialogBox dukeBox =
+                DialogBox.getDukeDialog("Hello! I'm Duke\nWhat can I do for you?", dukeImage);
+        dialogContainer.getChildren().addAll(dukeBox);
     }
 
     /**
