+++ src/main/java/duke/ui/Gui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 
 import javafx.application.Application;
 import javafx.scene.Scene;
