--- src/main/java/duke/ui/Ui.java
@@ -1,81 +1,0 @@
-package duke.ui;
-
-import duke.tasks.Task;
-
-/**
- * A class providing commonly used print functions.
- */
-public class Ui {
-
-    private static final String BAR =
-            "    ____________________________________________________________";
-
-    private static final String INDENTATION = "     ";
-    // private static final String NEW_LINE = "\n";
-
-    /**
-     * Prints an indented bar of line.
-     */
-    public static void printBar() {
-        System.out.println(BAR);
-    }
-
-    /**
-     * Prints {@code text} with indentation.
-     * 
-     * @param text
-     */
-    public static void printWithIndentation(String text) {
-        System.out.println(INDENTATION + text);
-    }
-
-    private static String getTextWithIndentation(String text) {
-        return INDENTATION + text;
-    }
-
-    /**
-     * Prints the welcome message for Duke.
-     */
-    public static void greet() {
-        prettyPrint("Hello! I'm Duke", "What can I do for you?");
-    }
-
-    /**
-     * Prints the newly added task.
-     * 
-     * @param task
-     * @param numTasks
-     */
-    public static void printAddTask(Task task, int numTasks) {
-        prettyPrint("Got it. I've added this task:", task.toString(),
-                "Now you have " + numTasks + " tasks in the list.");
-    }
-
-    public static String getAddTaskString(Task task, int numTasks) {
-        return getPrettyString("Got it. I've added this task:", task.toString(),
-                "Now you have " + numTasks + " tasks in the list.");
-    }
-
-    /**
-     * Prints the desired messages by indenting them and putting them in between bars.
-     * 
-     * @param textArray
-     */
-    public static void prettyPrint(String... textArray) {
-        printBar();
-        for (int i = 0; i < textArray.length; i++) {
-            printWithIndentation(textArray[i]);
-        }
-        printBar();
-    }
-
-    public static String getPrettyString(String... textArray) {
-        String prettyString = BAR;
-
-        for (int i = 0; i < textArray.length; i++) {
-            prettyString += getTextWithIndentation(textArray[i] + "\n");
-        }
-        prettyString += BAR;
-        return prettyString;
-    }
-}
